services/fixsuggest/requirements.txt
flask==2.3.2
pyyaml==6.0
gunicorn==20.1.0

flask-cors==4.0.0
